:recycle: refactor: #44 Home UI 컴포넌트 분리
상단 Banner 로 인해 전체가 렌더링되던 이슈
컴포넌트 분리로 개선
